\name{CaTCH-package}
\alias{CaTCH-package}
\alias{CaTCH}
\docType{package}
\title{
\packageTitle{CaTCH}
}
\description{
\packageDescription{CaTCH}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CaTCH}
\packageIndices{CaTCH}
}
\author{
\packageAuthor{CaTCH}

Maintainer: \packageMaintainer{CaTCH}
}
\references{
Zhan et al, Reciprocal insulation analysis of Hi-C data shows that TADs represent a functionally but not structurally privileged scale in the hierarchical folding of chromosomes, Genome Research 2017
}
\keyword{ TADs, domain calling, Hi-C }
\seealso{
}
\examples{
}
